requirements.txt
amqp==5.2.0
annotated-types==0.6.0
ansicon==1.89.0
anyio==4.3.0
arrow==1.3.0
asgiref==3.7.2
asttokens==2.4.1
attrs==23.2.0
billiard==4.2.0
blessed==1.20.0
celery==5.3.6
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
chartify==4.0.5
chartkick==1.0.1
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
comm==0.2.1
contextlib2==21.6.0
contourpy==1.2.0
cron-descriptor==1.4.3
cryptography==40.0.2
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
Django==5.0.2
django-allauth==0.54.0
django-background-tasks==1.2.5
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-compat==1.0.15
django-crontab==0.7.1
django-extensions==3.2.3
django-global-permissions==0.2.6
django-ninja==1.1.0
django-ninja-extra==0.20.4
django-ninja-jwt==5.3.0
django-picklefield==3.1
django-q==1.3.9
django-timezone-field==6.1.0
django-widget-tweaks==1.4.12
et-xmlfile==1.1.0
executing==2.0.1
flower==2.0.1
gunicorn==20.1.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
humanize==4.9.0
idna==3.4
injector==0.21.0
ipykernel==6.29.0
ipython==8.20.0
ipywidgets==8.1.1
jedi==0.19.1
Jinja2==3.1.3
jinxed==1.2.1
kombu==5.3.5
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mypy==1.3.0
mypy-extensions==1.0.0
mysql-connector==2.2.9
mysqlclient==2.2.1
nest-asyncio==1.5.9
oauthlib==3.2.2
openai==1.23.2
openpyxl==3.1.2
orjson==3.9.14
outcome==1.3.0.post0
packaging==23.1
parso==0.8.3
platformdirs==4.1.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
psutil==5.9.7
psycopg2-binary==2.9.9
pure-eval==0.2.2
pycparser==2.21
pydantic==2.6.1
pydantic_core==2.16.2
Pygments==2.17.2
PyJWT==2.7.0
PyMySQL==1.0.3
PySocks==1.7.1
python-crontab==3.0.0
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0.1

redis==3.5.3
requests==2.30.0
requests-oauthlib==1.3.1
selenium==4.16.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.4
stack-data==0.6.3
tenacity==8.2.2
tornado==6.4
tqdm==4.66.2
traitlets==5.14.1
trio==0.24.0
trio-websocket==0.11.1
types-python-dateutil==2.9.0.20240316
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.0.2
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.4.0
widgetsnbextension==4.0.9
wsproto==1.2.0
xyzservices==2023.10.1
